Update Navigation mixin
